Move contacts extract in CheckAuthCtrl
Extract is done in background.

Only contacts with email(s) AND phone(s) are send to backend.

Extract is not protected and will occurs each time CheckAuthCtrl
is called, it should be done only once per day (?) because it
seems time consuming (around 15 and 20 seconds ?)

POST is ok but I don't know if backend is ok or not